simplify api for RunQuery
